requirements.txt
fastapi>=0.104.1
uvicorn>=0.23.2
langchain>=0.1.0
langchain-core>=0.3.37
langgraph>=0.0.17
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
tavily-python>=0.2.8
openai>=1.3.7
langchain-openai>=0.0.2
nest-asyncio>=1.5.8
aiohttp>=3.9.1 